[IMP] sale_financial_risk: credit_limit change performance
res.partner credit_limit field is a commercial field that triggers the
children syncronization. In partners with lots of children, those
recalculations can lead to a poor user experience when a operation that
is replicated over all the children doesn't perform fine. This is the
case with the risk_partner_id field, wich relation is recalculated a
thus as many queries as children has the contact (it can be dozens).

TT43572